pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "API project using FastAPI, Qdrant, and MongoDB with Docker"
requires-python = ">=3.13"
authors = [
    { name = "Mathis LAMBERT", email = "mathislambert.dev@gmail.com" }
]
dependencies = [
    "fastapi[standard]>=0.115.8",
    "qdrant-client>=1.13.2",
    "motor>=3.7.0",
    "pymongo>=4.3.3",
    "uvicorn>=0.34.0",
    "mistralai>=1.5.0",
    "PyJWT>=2.10.1",
    "openai>=1.40.0",
    "httpx>=0.27.0",
    "anthropic>=0.39.0",
    "google-generativeai>=0.7.2",
    "pydantic-settings>=2.12.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0,<26.0.0",
    "flake8>=7.1.2,<7.2.0",
    "ruff>=0.0.283,<0.0.300",
    "isort>=5.12.0,<6.0.0",
    "mypy>=1.5.0,<1.6.0",
    "pytest-asyncio>=0.23.5"
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.ruff]
line-length = 88
exclude = [
    ".git",
    "__pycache__",
    ".mypy_cache",
    ".tox",
    ".venv",
    "build",
    "dist"
]

[tool.ruff.lint]
select = ["A", "B", "C", "E", "F", "I", "N", "Q", "T", "W"]
ignore = ["A002", "A003", "B008", "E203", "E501"]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.13"
warn_return_any = true
ignore_missing_imports = true
strict = true
